+++ src/series.rs
@@ -3,44 +3,44 @@
   for training the network.
 */
 
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 
 /// this macro improves the readability of the definition of all of the
 /// different series
 macro_rules! serie {
     ($name:ident ($($arg_name: ident: $arg_type: ty),*) $fun: expr) => {
-        pub fn $name($($arg_name: $arg_type),*) -> Box<dyn Fn(i32) -> f32> {
+        pub fn $name($($arg_name: $arg_type),*) -> Box<dyn Fn(i32) -> f64> {
             #![allow(unused)]
             Box::new($fun)
         }
     };
 }
 
-serie!(sine_with(period: i32, size: f32, offset: f32, phase: f32) move |i| {
-    (i as f32 * PI * 2.0 / period as f32 + phase).sin() * size + offset
+serie!(sine_with(period: i32, size: f64, offset: f64, phase: f64) move |i| {
+    (i as f64 * PI * 2.0 / period as f64 + phase).sin() * size + offset
 });
 
-serie!(saw_with(period: i32, size: f32, offset: f32, phase: f32) move |i| {
-    ((i as f32 / period as f32 + phase) % 1.0) * size - size / 2.0 + offset
+serie!(saw_with(period: i32, size: f64, offset: f64, phase: f64) move |i| {
+    ((i as f64 / period as f64 + phase) % 1.0) * size - size / 2.0 + offset
 });
 
-serie!(constant(value: f32) move |_| value);
+serie!(constant(value: f64) move |_| value);
 
-serie!(linear(period: i32, start: f32, end: f32) move |i| {
-    let t = i as f32 / period as f32;
+serie!(linear(period: i32, start: f64, end: f64) move |i| {
+    let t = i as f64 / period as f64;
     (1.0 - t) * start + t * end
 });
 
-serie!(spike(period: i32, height: f32, decay: f32) move |i| {
-    height * (- (i % period) as f32 / decay).exp()
+serie!(spike(period: i32, height: f64, decay: f64) move |i| {
+    height * (- (i % period) as f64 / decay).exp()
 });
 
-serie!(sine_speed_up(period: i32, size: f32, amount: f32, time: i32) move |i| {
-    let speedup = (1.0 + amount * (i as f32 / time as f32));
-    (i as f32 * PI * 2.0 * speedup / period as f32).sin() * size
+serie!(sine_speed_up(period: i32, size: f64, amount: f64, time: i32) move |i| {
+    let speedup = (1.0 + amount * (i as f64 / time as f64));
+    (i as f64 * PI * 2.0 * speedup / period as f64).sin() * size
 });
 
-serie!(impulse_pause(size: f32) move |i| {
+serie!(impulse_pause(size: f64) move |i| {
     if i == 0 {
         size
     } else {
@@ -48,7 +48,7 @@ serie!(impulse_pause(size: f32) move |i| {
     }
 });
 
-serie!(impulse_width_pause(size: f32, width: i32) move |i| {
+serie!(impulse_width_pause(size: f64, width: i32) move |i| {
     if i < width {
         size
     } else {
